fix(frontend): serve the archive at / instead of /copertine
The app owned `/` already but only to redirect it: `/` 307'd to
/copertine, so every visitor to copertine.ilmanifesto.it paid a
round-trip to land on a path that now merely repeats the domain name.

The prefix was never the app's own idea. On mema3 the archive shared a
vhost with a dozen unrelated services and /copertine was how nginx told
them apart. mema4 gave it its own host and its own domain, which is what
made the segment redundant — the move to Dokploy collapsed five nginx
location blocks into two Traefik routes, and this is the last piece of
that collapse.

- app/copertine/page.tsx -> app/page.tsx, replacing the redirect stub.
  Relative imports shift up one level (../ -> ./).
- app/copertine/layout.tsx is folded into the root layout. Its Suspense
  boundary is load-bearing, not decoration: the page is a client
  component reading searchParams, which Next requires be wrapped or the
  build fails.
- /copertine now redirects to / permanently, so links held externally
  from the mema3 era keep working. /api/copertine is a different route
  and is untouched.

Verified against a production build: / serves 200 with the header and
Suspense fallback rendered, /copertine 308s to /, and the route table
no longer lists a /copertine page.
