+++ frontend/app/layout.tsx
@@ -1,7 +1,9 @@
 // app/layout.tsx
+import { Suspense } from "react";
 import type { Metadata } from "next";
 import { Inter as FontSans } from "next/font/google";
 import { ThemeProvider } from "../providers/theme-provider";
+import Header from "./components/header/Header";
 import "./globals.css";
 
 const fontSans = FontSans({
@@ -31,7 +33,26 @@ export default function RootLayout({
       </head>
       <body className="antialiased">
         <ThemeProvider defaultTheme="light">
-          {children}
+          {/* Header and the Suspense boundary used to live in a nested
+              app/copertine/layout.tsx. The archive now owns `/` outright, so
+              there is no nested segment left to hold them and they belong
+              here. The boundary is load-bearing, not decoration: the page is
+              a client component that reads searchParams, which Next requires
+              to be wrapped in Suspense or the build fails. */}
+          <Header />
+          <main>
+            <Suspense
+              fallback={
+                <div className="min-h-[calc(100vh-176px)] flex flex-col items-center justify-center">
+                  <div className="animate-pulse text-blue-700 text-lg">
+                    Caricamento...
+                  </div>
+                </div>
+              }
+            >
+              {children}
+            </Suspense>
+          </main>
         </ThemeProvider>
       </body>
     </html>
