+++ frontend/next.config.ts
@@ -8,6 +8,17 @@ const nextConfig: NextConfig = {
     unoptimized: true,
     domains: ['localhost'],
   },
+  async redirects() {
+    return [
+      // The archive lived at /copertine for years — as a subpath of the shared
+      // mema3 vhost, and then briefly on its own host, where `/` merely
+      // redirected here. Now that it owns `/` the prefix is redundant with the
+      // domain name, but externally held links still carry it. Redirect the
+      // other way so those keep working. Permanent: the old path is not coming
+      // back. /api/copertine is a different route and is untouched.
+      { source: "/copertine", destination: "/", permanent: true },
+    ];
+  },
 };
 
 export default nextConfig;
